Extract native CLI install path resolution
Move install path calculation, comparison, ownership checks, and environment reads into a focused resolver. Keep current-process PATH mutation in the installer so value resolution remains separate from state changes.
