+++ Packages/src/Editor/Infrastructure/CLI/NativeCliInstallPathResolver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f0435a28178b4125b2b9a1eb26360be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
