+++ Packages/src/Editor/Infrastructure/CLI/NativeCliInstaller.cs
@@ -20,8 +20,6 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
         private const int INSTALL_PROCESS_TIMEOUT_MS = 300000;
         private const int INSTALL_PROCESS_WAIT_SLICE_MS = 250;
         internal const int UNINSTALL_COMPLETION_TIMEOUT_MS = 30000;
-        private const string WINDOWS_FILE_PATH_SEPARATOR = "\\";
-        private const string POSIX_FILE_PATH_SEPARATOR = "/";
 
         public static NativeCliInstallCommand GetInstallCommand(
             RuntimePlatform platform,
@@ -44,7 +42,7 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
             UnityEngine.Debug.Assert(!string.IsNullOrWhiteSpace(cliReleaseTag), "cliReleaseTag must not be null or empty");
             ct.ThrowIfCancellationRequested();
 
-            string installDirectory = GetInstallDirectoryForCurrentUser(platform);
+            string installDirectory = NativeCliInstallPathResolver.GetInstallDirectoryForCurrentUser(platform);
             if (string.IsNullOrWhiteSpace(installDirectory))
             {
                 return new CliInstallResult(
@@ -73,7 +71,7 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
         {
             ct.ThrowIfCancellationRequested();
 
-            string installDirectory = GetInstallDirectoryForCurrentUser(platform);
+            string installDirectory = NativeCliInstallPathResolver.GetInstallDirectoryForCurrentUser(platform);
             if (string.IsNullOrWhiteSpace(installDirectory))
             {
                 return new CliInstallResult(
@@ -90,7 +88,7 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
                 return result;
             }
 
-            string installPath = GetGlobalCliInstallPath(installDirectory, platform);
+            string installPath = NativeCliInstallPathResolver.GetGlobalCliInstallPath(installDirectory, platform);
             CliInstallResult removalResult = await WaitForUninstallCompletionAsync(
                 installPath,
                 installDirectory,
@@ -117,7 +115,7 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
         {
             UnityEngine.Debug.Assert(!string.IsNullOrWhiteSpace(installDirectory), "installDirectory must not be null or empty");
 
-            string installPath = GetGlobalCliInstallPath(installDirectory, platform);
+            string installPath = NativeCliInstallPathResolver.GetGlobalCliInstallPath(installDirectory, platform);
             return new NativeCliInstallCommand(
                 installPath,
                 "uninstall",
@@ -336,15 +334,6 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
             }
         }
 
-        internal static string GetGlobalCliInstallPath(string installDirectory, RuntimePlatform platform)
-        {
-            UnityEngine.Debug.Assert(!string.IsNullOrEmpty(installDirectory), "installDirectory must not be null or empty");
-
-            string separator = GetFilePathSeparator(platform);
-            string normalizedInstallDirectory = installDirectory.TrimEnd('\\', '/');
-            return normalizedInstallDirectory + separator + GetGlobalCliInstallFileName(platform);
-        }
-
         internal static CliInstallResult FinishSuccessfulInstall(
             CliInstallResult installResult,
             string installDirectory,
@@ -359,274 +348,38 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
             return installResult;
         }
 
-        internal static string BuildPathWithInstallDirectory(
-            string currentPath,
-            string installDirectory,
-            RuntimePlatform platform)
-        {
-            UnityEngine.Debug.Assert(!string.IsNullOrWhiteSpace(installDirectory), "installDirectory must not be null or empty");
-
-            string normalizedPath = currentPath ?? "";
-            if (string.IsNullOrEmpty(normalizedPath))
-            {
-                return installDirectory;
-            }
-
-            string separator = GetPathSeparator(platform);
-            string[] entries = normalizedPath.Split(
-                new[] { separator },
-                StringSplitOptions.RemoveEmptyEntries);
-            StringComparison comparison = GetPathComparison(platform);
-            StringBuilder builder = new(installDirectory);
-            foreach (string entry in entries)
-            {
-                if (string.Equals(entry, installDirectory, comparison))
-                {
-                    continue;
-                }
-
-                builder.Append(separator);
-                builder.Append(entry);
-            }
-
-            return builder.ToString();
-        }
-
-        internal static string BuildPathWithoutInstallDirectory(
-            string currentPath,
-            string installDirectory,
-            RuntimePlatform platform)
-        {
-            UnityEngine.Debug.Assert(!string.IsNullOrWhiteSpace(installDirectory), "installDirectory must not be null or empty");
-
-            string normalizedPath = currentPath ?? "";
-            if (string.IsNullOrEmpty(normalizedPath))
-            {
-                return "";
-            }
-
-            string separator = GetPathSeparator(platform);
-            string[] entries = normalizedPath.Split(
-                new[] { separator },
-                StringSplitOptions.RemoveEmptyEntries);
-            string normalizedInstallDirectory = NormalizePathForComparison(installDirectory, platform);
-            StringComparison comparison = GetPathComparison(platform);
-            StringBuilder builder = new StringBuilder();
-            foreach (string entry in entries)
-            {
-                if (string.IsNullOrWhiteSpace(entry))
-                {
-                    continue;
-                }
-
-                string normalizedEntry = NormalizePathForComparison(entry, platform);
-                if (string.Equals(normalizedEntry, normalizedInstallDirectory, comparison))
-                {
-                    continue;
-                }
-
-                if (builder.Length > 0)
-                {
-                    builder.Append(separator);
-                }
-
-                builder.Append(entry);
-            }
-
-            return builder.ToString();
-        }
-
-        internal static string GetDefaultInstallDirectoryFromRoots(
-            RuntimePlatform platform,
-            string homeDirectory,
-            string localAppData)
-        {
-            if (platform == RuntimePlatform.WindowsEditor)
-            {
-                if (string.IsNullOrWhiteSpace(localAppData))
-                {
-                    return null;
-                }
-
-                return Path.Combine(
-                    localAppData,
-                    CliConstants.WINDOWS_PROGRAMS_DIR_NAME,
-                    CliConstants.NATIVE_INSTALL_DIR_NAME,
-                    CliConstants.NATIVE_INSTALL_BIN_DIR_NAME);
-            }
-
-            if (string.IsNullOrWhiteSpace(homeDirectory))
-            {
-                return null;
-            }
-
-            return Path.Combine(
-                homeDirectory,
-                CliConstants.POSIX_LOCAL_DIR_NAME,
-                CliConstants.NATIVE_INSTALL_BIN_DIR_NAME);
-        }
-
         private static void ApplyInstallDirectoryToCurrentProcessPath(RuntimePlatform platform)
         {
-            string installDirectory = GetInstallDirectoryForCurrentUser(platform);
+            string installDirectory = NativeCliInstallPathResolver.GetInstallDirectoryForCurrentUser(platform);
             if (string.IsNullOrEmpty(installDirectory))
             {
                 return;
             }
 
-            string pathVariableName = GetPathEnvironmentVariableName(platform);
+            string pathVariableName = NativeCliInstallPathResolver.GetPathEnvironmentVariableName(platform);
             string currentPath = Environment.GetEnvironmentVariable(pathVariableName);
-            string updatedPath = BuildPathWithInstallDirectory(currentPath, installDirectory, platform);
+            string updatedPath = NativeCliInstallPathResolver.BuildPathWithInstallDirectory(
+                currentPath,
+                installDirectory,
+                platform);
             Environment.SetEnvironmentVariable(pathVariableName, updatedPath);
         }
 
         private static void RemoveInstallDirectoryFromCurrentProcessPath(RuntimePlatform platform)
         {
-            string installDirectory = GetInstallDirectoryForCurrentUser(platform);
+            string installDirectory = NativeCliInstallPathResolver.GetInstallDirectoryForCurrentUser(platform);
             if (string.IsNullOrEmpty(installDirectory))
             {
                 return;
             }
 
-            string pathVariableName = GetPathEnvironmentVariableName(platform);
+            string pathVariableName = NativeCliInstallPathResolver.GetPathEnvironmentVariableName(platform);
             string currentPath = Environment.GetEnvironmentVariable(pathVariableName);
-            string updatedPath = BuildPathWithoutInstallDirectory(currentPath, installDirectory, platform);
+            string updatedPath = NativeCliInstallPathResolver.BuildPathWithoutInstallDirectory(
+                currentPath,
+                installDirectory,
+                platform);
             Environment.SetEnvironmentVariable(pathVariableName, updatedPath);
-        }
-
-        internal static bool IsPackageOwnedCurrentUserInstallPath(
-            string executablePath,
-            RuntimePlatform platform)
-        {
-            if (string.IsNullOrWhiteSpace(executablePath))
-            {
-                return false;
-            }
-
-            string installDirectory = GetInstallDirectoryForCurrentUser(platform);
-            if (string.IsNullOrWhiteSpace(installDirectory))
-            {
-                return false;
-            }
-
-            return IsPackageOwnedInstallPath(executablePath, installDirectory, platform);
-        }
-
-        internal static bool IsPackageOwnedInstallPath(
-            string executablePath,
-            string installDirectory,
-            RuntimePlatform platform)
-        {
-            UnityEngine.Debug.Assert(!string.IsNullOrWhiteSpace(installDirectory), "installDirectory must not be null or empty");
-
-            if (string.IsNullOrWhiteSpace(executablePath))
-            {
-                return false;
-            }
-
-            string expectedPath = GetGlobalCliInstallPath(installDirectory, platform);
-            string normalizedExecutablePath = NormalizePathForComparison(executablePath, platform);
-            string normalizedExpectedPath = NormalizePathForComparison(expectedPath, platform);
-            return string.Equals(
-                normalizedExecutablePath,
-                normalizedExpectedPath,
-                GetPathComparison(platform));
-        }
-
-        internal static string GetCurrentUserGlobalCliInstallPath(RuntimePlatform platform)
-        {
-            string installDirectory = GetInstallDirectoryForCurrentUser(platform);
-            return string.IsNullOrWhiteSpace(installDirectory)
-                ? null
-                : GetGlobalCliInstallPath(installDirectory, platform);
-        }
-
-        internal static string GetCurrentUserGlobalCliInstallDirectory(RuntimePlatform platform)
-        {
-            return GetInstallDirectoryForCurrentUser(platform);
-        }
-
-        internal static bool HasPackageOwnedCurrentUserInstall(RuntimePlatform platform)
-        {
-            string executablePath = GetCurrentUserGlobalCliInstallPath(platform);
-            return !string.IsNullOrWhiteSpace(executablePath) && File.Exists(executablePath);
-        }
-
-        private static string GetInstallDirectoryForCurrentUser(RuntimePlatform platform)
-        {
-            string configuredInstallDirectory = Environment.GetEnvironmentVariable(CliConstants.INSTALL_DIR_ENVIRONMENT_VARIABLE);
-            if (!string.IsNullOrWhiteSpace(configuredInstallDirectory))
-            {
-                return configuredInstallDirectory;
-            }
-
-            string homeDirectory = Environment.GetEnvironmentVariable(CliConstants.POSIX_HOME_ENVIRONMENT_VARIABLE);
-            if (string.IsNullOrWhiteSpace(homeDirectory))
-            {
-                homeDirectory = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-            }
-
-            string localAppData = Environment.GetEnvironmentVariable(CliConstants.WINDOWS_LOCAL_APPDATA_ENVIRONMENT_VARIABLE);
-            return GetDefaultInstallDirectoryFromRoots(platform, homeDirectory, localAppData);
-        }
-
-        private static string GetPathEnvironmentVariableName(RuntimePlatform platform)
-        {
-            return platform == RuntimePlatform.WindowsEditor
-                ? CliConstants.WINDOWS_PATH_ENVIRONMENT_VARIABLE
-                : CliConstants.POSIX_PATH_ENVIRONMENT_VARIABLE;
-        }
-
-        private static string GetPathSeparator(RuntimePlatform platform)
-        {
-            return platform == RuntimePlatform.WindowsEditor
-                ? CliConstants.WINDOWS_PATH_SEPARATOR
-                : CliConstants.POSIX_PATH_SEPARATOR;
-        }
-
-        private static string GetFilePathSeparator(RuntimePlatform platform)
-        {
-            return platform == RuntimePlatform.WindowsEditor
-                ? WINDOWS_FILE_PATH_SEPARATOR
-                : POSIX_FILE_PATH_SEPARATOR;
-        }
-
-        private static StringComparison GetPathComparison(RuntimePlatform platform)
-        {
-            return platform == RuntimePlatform.WindowsEditor
-                ? StringComparison.OrdinalIgnoreCase
-                : StringComparison.Ordinal;
-        }
-
-        private static string NormalizePathForComparison(string path, RuntimePlatform platform)
-        {
-            UnityEngine.Debug.Assert(!string.IsNullOrWhiteSpace(path), "path must not be null or empty");
-
-            string normalizedPath = path.Trim().Trim('"');
-            if (platform != RuntimePlatform.WindowsEditor)
-            {
-                return normalizedPath.TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
-            }
-
-            return normalizedPath.TrimEnd('\\', '/').Replace('/', '\\');
-        }
-
-        private static bool DoesUserPathContainInstallDirectory(
-            string installDirectory,
-            RuntimePlatform platform,
-            Func<string, EnvironmentVariableTarget, string> getEnvironmentVariable)
-        {
-            UnityEngine.Debug.Assert(!string.IsNullOrWhiteSpace(installDirectory), "installDirectory must not be null or empty");
-            UnityEngine.Debug.Assert(getEnvironmentVariable != null, "getEnvironmentVariable must not be null");
-
-            if (platform != RuntimePlatform.WindowsEditor)
-            {
-                return false;
-            }
-
-            string pathVariableName = GetPathEnvironmentVariableName(platform);
-            string currentUserPath = getEnvironmentVariable(pathVariableName, EnvironmentVariableTarget.User);
-            return DoesPathContainInstallDirectory(currentUserPath, installDirectory, platform);
         }
 
         private static bool ShouldWaitForUserPathRemoval(
@@ -639,43 +392,10 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
             UnityEngine.Debug.Assert(getEnvironmentVariable != null, "getEnvironmentVariable must not be null");
 
             return requireUserPathRemoval
-                && DoesUserPathContainInstallDirectory(installDirectory, platform, getEnvironmentVariable);
-        }
-
-        private static bool DoesPathContainInstallDirectory(
-            string currentPath,
-            string installDirectory,
-            RuntimePlatform platform)
-        {
-            UnityEngine.Debug.Assert(!string.IsNullOrWhiteSpace(installDirectory), "installDirectory must not be null or empty");
-
-            string normalizedPath = currentPath ?? "";
-            if (string.IsNullOrEmpty(normalizedPath))
-            {
-                return false;
-            }
-
-            string separator = GetPathSeparator(platform);
-            string[] entries = normalizedPath.Split(
-                new[] { separator },
-                StringSplitOptions.RemoveEmptyEntries);
-            string normalizedInstallDirectory = NormalizePathForComparison(installDirectory, platform);
-            StringComparison comparison = GetPathComparison(platform);
-            foreach (string entry in entries)
-            {
-                if (string.IsNullOrWhiteSpace(entry))
-                {
-                    continue;
-                }
-
-                string normalizedEntry = NormalizePathForComparison(entry, platform);
-                if (string.Equals(normalizedEntry, normalizedInstallDirectory, comparison))
-                {
-                    return true;
-                }
-            }
-
-            return false;
+                && NativeCliInstallPathResolver.DoesUserPathContainInstallDirectory(
+                    installDirectory,
+                    platform,
+                    getEnvironmentVariable);
         }
 
         private static string BuildUninstallCompletionTimeoutFailure(
@@ -797,13 +517,6 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
             }
         }
 
-        private static string GetGlobalCliInstallFileName(RuntimePlatform platform)
-        {
-            return platform == RuntimePlatform.WindowsEditor
-                ? CliConstants.GLOBAL_WINDOWS_COMMAND_NAME
-                : CliConstants.GLOBAL_UNIX_COMMAND_NAME;
-        }
-
     }
 
     /// <summary>
@@ -813,12 +526,12 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
     {
         public bool IsPackageOwnedCurrentUserInstallPath(string cliExecutablePath, RuntimePlatform platform)
         {
-            return NativeCliInstaller.IsPackageOwnedCurrentUserInstallPath(cliExecutablePath, platform);
+            return NativeCliInstallPathResolver.IsPackageOwnedCurrentUserInstallPath(cliExecutablePath, platform);
         }
 
         public bool HasPackageOwnedCurrentUserInstall(RuntimePlatform platform)
         {
-            return NativeCliInstaller.HasPackageOwnedCurrentUserInstall(platform);
+            return NativeCliInstallPathResolver.HasPackageOwnedCurrentUserInstall(platform);
         }
 
         public Task<CliInstallResult> InstallGlobalCliAsync(
